# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hadaya Gift Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - in: query
          name: collectionId
          schema: { type: ["integer", "null"] }
        - in: query
          name: search
          schema: { type: ["string", "null"] }
        - in: query
          name: featured
          schema: { type: ["boolean", "null"] }
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  # ─── Customer: Collections ────────────────────────────────────────────────
...
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get collection with its products
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Collection with products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionWithProducts"
        "404":
          description: Not found

  # ─── Customer: Orders ─────────────────────────────────────────────────────
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  # ─── Gift Flow ────────────────────────────────────────────────────────────
  /gift/{token}:
    get:
      operationId: getGiftByToken
      tags: [gift]
      summary: Get gift order info by secure token
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Gift info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftInfo"
        "404":
          description: Token not found or expired
  /gift/{token}/address:
    post:
      operationId: submitGiftAddress
      tags: [gift]
      summary: Recipient submits their delivery address
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftAddressInput"
      responses:
        "200":
          description: Address saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftInfo"
        "404":
          description: Token not found or expired

  # ─── Admin: Dashboard Stats ───────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, nameAr, nameEn, price, imageUrl, inStock, isFeatured, createdAt]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: ["string", "null"] }
        descriptionEn: { type: ["string", "null"] }
        price: { type: number }
        imageUrl: { type: string }
        inStock: { type: boolean }
        isFeatured: { type: boolean }
        collectionId: { type: ["integer", "null"] }
        collectionName: { type: ["string", "null"] }
        createdAt: { type: string }
...
    CollectionWithProducts:
      type: object
      required: [id, nameAr, nameEn, products, createdAt]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: ["string", "null"] }
        descriptionEn: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        createdAt: { type: string }
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
...
    OrderItem:
      type: object
      required: [id, productId, productNameAr, productNameEn, productImageUrl, quantity, unitPrice, subtotal]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productNameAr: { type: string }
        productNameEn: { type: string }
        productImageUrl: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        subtotal: { type: number }
    Order:
      type: object
      required: [id, status, paymentMethod, subtotal, deliveryFee, total, createdAt, items]
      properties:
        id: { type: integer }
        customerName: { type: ["string", "null"] }
        customerPhone: { type: ["string", "null"] }
        customerEmail: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        paymentMethod:
          type: string
          enum: [cash_on_delivery, bank_transfer, card]
        subtotal: { type: number }
        deliveryFee: { type: number }
        total: { type: number }
        notes: { type: ["string", "null"] }
        recipientAddress: { type: ["string", "null"] }
        recipientCity: { type: ["string", "null"] }
        giftToken: { type: ["string", "null"] }
        isGift: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderInput:
      type: object
      required: [items, paymentMethod]
      properties:
        customerName: { type: string }
        customerPhone: { type: string }
        customerEmail: { type: string }
        paymentMethod:
          type: string
          enum: [cash_on_delivery, bank_transfer, card]
        notes: { type: string }
        isGift: { type: boolean }
        recipientAddress: { type: string }
        recipientCity: { type: string }
        items:
          type: array
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId: { type: integer }
              quantity: { type: integer }
...
    GiftInfo:
      type: object
      required: [orderId, isAddressSubmitted, items]
      properties:
        orderId: { type: integer }
        senderName: { type: ["string", "null"] }
        message: { type: ["string", "null"] }
        isAddressSubmitted: { type: boolean }
        recipientAddress: { type: ["string", "null"] }
        recipientCity: { type: ["string", "null"] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    GiftAddressInput:
      type: object
      required: [recipientAddress, recipientCity]
      properties:
        recipientAddress: { type: string, minLength: 1 }
        recipientCity: { type: string, minLength: 1 }
...