# lib/api-zod/src/generated/types/collectionWithProducts.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hadaya Gift Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CollectionWithProducts {
  id: number;
  nameAr: string;
  nameEn: string;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  products: Product[];
}