# lib/api-zod/src/generated/types/giftInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hadaya Gift Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface GiftInfo {
  orderId: number;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  message?: string | null;
  isAddressSubmitted: boolean;
  /** @nullable */
  recipientAddress?: string | null;
  /** @nullable */
  recipientCity?: string | null;
  items: OrderItem[];
}